requirements.txt
Django==2.0.3
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
gunicorn==19.7.1
PyJWT==1.6.1
pytz==2018.3

